+++ tests/Integration/Framework/UsingConfigInterfacesMapping/LocalConfig/Domain/GreeterGeneratorInterface.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace GacelaTest\Integration\Framework\UsingConfigDependencies\LocalConfig\Domain;
+namespace GacelaTest\Integration\Framework\UsingConfigInterfacesMapping\LocalConfig\Domain;
 
 interface GreeterGeneratorInterface
 {