+++ tests/Integration/Framework/UsingConfigInterfacesMapping/LocalConfig/Domain/NumberService.php
@@ -2,9 +2,9 @@
 
 declare(strict_types=1);
 
-namespace GacelaTest\Integration\Framework\UsingConfigDependencies\LocalConfig\Domain;
+namespace GacelaTest\Integration\Framework\UsingConfigInterfacesMapping\LocalConfig\Domain;
 
-use GacelaTest\Integration\Framework\UsingConfigDependencies\LocalConfig\Infrastructure\CustomCompanyGenerator;
+use GacelaTest\Integration\Framework\UsingConfigInterfacesMapping\LocalConfig\Infrastructure\CustomCompanyGenerator;
 
 final class NumberService
 {