+++ tests/Integration/Framework/UsingConfigInterfacesMapping/LocalConfig/Factory.php
@@ -2,11 +2,11 @@
 
 declare(strict_types=1);
 
-namespace GacelaTest\Integration\Framework\UsingConfigDependencies\LocalConfig;
+namespace GacelaTest\Integration\Framework\UsingConfigInterfacesMapping\LocalConfig;
 
 use Gacela\Framework\AbstractFactory;
-use GacelaTest\Integration\Framework\UsingConfigDependencies\LocalConfig\Domain\GreeterGeneratorInterface;
-use GacelaTest\Integration\Framework\UsingConfigDependencies\LocalConfig\Domain\NumberService;
+use GacelaTest\Integration\Framework\UsingConfigInterfacesMapping\LocalConfig\Domain\GreeterGeneratorInterface;
+use GacelaTest\Integration\Framework\UsingConfigInterfacesMapping\LocalConfig\Domain\NumberService;
 
 final class Factory extends AbstractFactory
 {