+++ tests/Integration/Framework/UsingConfigInterfacesMapping/LocalConfig/Infrastructure/CustomCompanyGenerator.php
@@ -2,9 +2,9 @@
 
 declare(strict_types=1);
 
-namespace GacelaTest\Integration\Framework\UsingConfigDependencies\LocalConfig\Infrastructure;
+namespace GacelaTest\Integration\Framework\UsingConfigInterfacesMapping\LocalConfig\Infrastructure;
 
-use GacelaTest\Integration\Framework\UsingConfigDependencies\LocalConfig\Domain\GreeterGeneratorInterface;
+use GacelaTest\Integration\Framework\UsingConfigInterfacesMapping\LocalConfig\Domain\GreeterGeneratorInterface;
 
 final class CustomCompanyGenerator implements GreeterGeneratorInterface
 {