+++ tests/Integration/Framework/UsingConfigInterfacesMapping/gacela.php
@@ -1,0 +1,12 @@
+<?php
+
+declare(strict_types=1);
+
+use GacelaTest\Integration\Framework\UsingConfigInterfacesMapping\LocalConfig\Domain\GreeterGeneratorInterface;
+use GacelaTest\Integration\Framework\UsingConfigInterfacesMapping\LocalConfig\Infrastructure\CustomCompanyGenerator;
+
+return [
+    'interfaces-mapping' => [
+        GreeterGeneratorInterface::class => CustomCompanyGenerator::class,
+    ],
+];